**IMPORTANT UPDATE**
In doing some testing, Emulator.JS uses an older version of dosbox pure which allows the following configs to work (using DOOM as an example):

1. dosbox.conf in the root of the zip
2. doom.conf inside the zip in the same location as doom.exe
3. doom.conf in the same *folders* as the zip, but not inside.

Current testing with dosbox-pure cores (pre3) show that only the first one works. I've updated the application with a checkbox that allows you to insert the properly named conf at the root of the zip.